the products will now populate on the main page

+++ server/seeders/cleanDB.js
@@ -1,9 +1,17 @@
 const models = require("../models");
 const db = require("../config/connection");
 
-module.exports = async (collectionName) => {
+module.exports = async (modelName, collectionName) => {
   try {
-    await db.dropCollection(collectionName);
+    let modelExists = await models[modelName].db.db
+      .listCollections({
+        name: collectionName,
+      })
+      .toArray();
+
+    if (modelExists.length) {
+      await db.dropCollection(collectionName);
+    }
   } catch (err) {
     throw err;
   }
